Fix minor logic error in hawtio proxy script
If no argument was provided on invocation, the $port_forward_pid
variable would be unset.  The cleanup function would try to run "ps -p"
on that variable as part of the EXIT trap.  A little error message about
a missing argument to the "-p" would then print.

+++ bin/hawtio-proxy.sh
@@ -8,7 +8,7 @@ nginx_config=$(mktemp --suffix=.conf --tmpdir nginx-XXXXXX)
 function cleanup() {
   rm -f "$nginx_config"
 
-  if ps -p "$port_forward_pid" > /dev/null; then
+  if [[ -v $port_forward_pid ]] && ps -p "$port_forward_pid" > /dev/null; then
     # The - signifies we should kill all processes in the process group
     kill -- -$port_forward_pid
   fi
